ft_strlen.c returns size_t

+++ ft_strlen.c
@@ -6,11 +6,11 @@
 /*   By: tmentor <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/20 12:17:15 by tmentor           #+#    #+#             */
-/*   Updated: 2019/05/20 12:23:57 by tmentor          ###   ########.fr       */
+/*   Updated: 2019/05/21 10:11:04 by tmentor          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int		ft_strlen(char *str)
+size_t		ft_strlen(char *str)
 {
 	int i;
